pyproject.toml
[project]
authors = [{ name = "Tang Tao", email = "2024111155@stu.sufe.edu.cn" }]
description = "A machine learning framework project."
name = "ml-framework"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytorch-ignite>=0.5.3",
    "tensorboard>=2.20.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.ruff]

line-length = 120
